modernizacion/backend/requirements.txt
requests
flask
psycopg2-binary
flask-sqlalchemy
marshmallow-sqlalchemy
marshmallow
flask-marshmallow
flask-cors
python-dotenv
pytest
pytest-cov~=4.0.0
pytest-mock~=3.10.0
pytest-env~=0.8.1
aws-wsgi==0.2.7